+++ qcs-api-client-openapi/internal/src/models/drag_gaussian_waveform.rs
@@ -16,38 +16,38 @@ pub struct DragGaussianWaveform {
     #[serde(rename = "_type")]
     pub _type: String,
     #[serde(rename = "alpha")]
-    pub alpha: f32,
+    pub alpha: f64,
     #[serde(rename = "anh")]
-    pub anh: f32,
+    pub anh: f64,
     #[serde(rename = "detuning")]
-    pub detuning: f32,
+    pub detuning: f64,
     #[serde(rename = "duration")]
-    pub duration: f32,
+    pub duration: f64,
     #[serde(rename = "frame")]
     pub frame: String,
     #[serde(rename = "fwhm")]
-    pub fwhm: f32,
+    pub fwhm: f64,
     #[serde(rename = "phase")]
-    pub phase: f32,
+    pub phase: f64,
     #[serde(rename = "scale")]
-    pub scale: f32,
+    pub scale: f64,
     #[serde(rename = "t0")]
-    pub t0: f32,
+    pub t0: f64,
 }
 
 impl DragGaussianWaveform {
     /// Gaussian-like waveform used for transmons to correct for the transmon's anharmonicity.  DRAG === Derivative Removal by Adiabatic Gate.  Reference: Motzoi, et al, Simple Pulses for Elimination of Leakage in Weakly Nonlinear Qubits. DOI: 10.1103/PhysRevLett.103.110501
     pub fn new(
         _type: String,
-        alpha: f32,
-        anh: f32,
-        detuning: f32,
-        duration: f32,
+        alpha: f64,
+        anh: f64,
+        detuning: f64,
+        duration: f64,
         frame: String,
-        fwhm: f32,
-        phase: f32,
-        scale: f32,
-        t0: f32,
+        fwhm: f64,
+        phase: f64,
+        scale: f64,
+        t0: f64,
     ) -> DragGaussianWaveform {
         DragGaussianWaveform {
             _type,
