+++ qcs-api-client-openapi/internal/src/models/qdo_fast_flux_channel.rs
@@ -18,11 +18,11 @@ pub struct QdoFastFluxChannel {
     #[serde(rename = "channel_index")]
     pub channel_index: i32,
     #[serde(rename = "delay", skip_serializing_if = "Option::is_none")]
-    pub delay: Option<f32>,
+    pub delay: Option<f64>,
     #[serde(rename = "direction", skip_serializing_if = "Option::is_none")]
     pub direction: Option<Box<crate::models::PropagationDirection>>,
     #[serde(rename = "flux_current", skip_serializing_if = "Option::is_none")]
-    pub flux_current: Option<f32>,
+    pub flux_current: Option<f64>,
 }
 
 impl QdoFastFluxChannel {
